Crate lojidoc

Re-exports

extern crate std as std;

extern crate clap;

extern crate regex;

extern crate threadpool;

use std::prelude::v1::*;

use clap::App;

use clap::Arg;

use std::fs;

use std::fs::File;

use std::io::prelude::*;

use std::path::Path;

use std::path::PathBuf;

use std::sync::Arc;

use threadpool::ThreadPool;

use model::model::Class;

use model::model::Method;

use model::model::Interface;

use model::model::Project;

use model::model::LineType;

use parse::parse::parse_file;

Modules

model

parse

Functions

document
Handles the thread pooling the application

document_single
Handles the single threaded option for running the application

find_java_files
Traverses the file structure to find all java files for parsing.

find_repo_home
Finds the root directory of the cloned repository

gen_class_docs
Generates the markdown documentation for a class

gen_interface_docs
Generates the markdown documentation for an interface

gen_method_docs
Generates the markdown documentation for the methods of a class

generate_markdown
Generates a markdown file for a java file Uses a Class struct to write the markdown

is_java_file

is_repo_dir
Determines whether a file path contains a git or mercurial file

lint_javadoc
Handles linting javadocs without saving the documentation

main

resolve_context
Combines the repo url with java file path to provide a link in the docs