Module lojidoc::document::document

Re-exports

extern crate colored;

use mdbook::MDBook;

use std::fs;

use std::fs::File;

use std::io::prelude::*;

use std::path::Path;

use std::path::PathBuf;

use colored::*;

use model::contents::ApplicationDoc;

use model::model::Class;

use model::model::Interface;

use model::model::Member;

use model::model::Method;

use model::model::Project;

Functions

find_gen_files
Traverses the file structure to find all java files for parsing.

find_java_files
Traverses the file structure to find all java files for parsing.

find_repo_home
Finds the root directory of the cloned repository

gen_class_docs
Generates the markdown documentation for a class

gen_interface_docs
Generates the markdown documentation for an interface

gen_md_book
Creates a markdown book using mdbook. Uses the files in the generated documentation directory for chapters.

gen_method_docs
Generates the markdown documentation for the methods of a class

gen_var_docs
Generates the markdown documentation for the member variables of a class

generate_markdown
Generates a markdown file for a java file Uses a Class struct to write the markdown

is_java_file
Finds out whether a file is a java file

is_repo_dir
Determines whether a file path contains a git or mercurial file

lint_method

lint_project
Lints the java project's javadoc comments and prints the errors

lint_var

resolve_context
Combines the repo url with java file path to provide a link in the docs