Module lojidoc::grammar::grammar

Module that contains grammar used in the lexing and parsing of java code Also defines other sets of keywords like javadoc keywords or framework annotations

Structs

ParseState
Struct that represents the parsing state of the high level java declarations

Enums

JdocState
Stores the state of javadoc parsing. Each enum field represents a javadoc keyword defined in the get_jdoc_keywords() function.

JdocToken

Stream

Token

Functions

get_jdoc_keywords
Gets a full list of all the javadoc keywords for the lexer

get_keywords
Gets a full list of all the keywords for the lexer