Module lojidoc::parse::parse

A module which handles the parsing for java files

Re-exports

extern crate regex;

use model::model::*;

use regex::Regex;

use std::fs::File;

use std::io::BufRead;

use std::io::BufReader;

use std::path::Path;

Functions

determine_line_type
Determines the line type of a line of java code

doc_desc

end_doc_match

handle_class

handle_doc
Handles parsing javadoc comments

handle_interface

handle_method

match_params

parse_file

regex_match

start_comment_match

start_doc_match

trim_paren

trim_whitespace
Removes all whitespcace from a line