Module lojidoc::parse::parse

A module which handles the parsing for java files

Re-exports

use grammar::grammar::*;

use model::model::Class;

use model::model::Doc;

use model::model::Exception;

use model::model::Member;

use model::model::Method;

use model::model::Object;

use model::model::ObjectState;

use model::model::ObjectType;

use model::model::Param;

use colored::*;

use std::fs::File;

use std::io::BufReader;

use std::io::Read;

use std::path::Path;

Functions

construct_ast
Constucts a syntax tree based on the stream of token from the lexing Outputs a Class struct containing all the data for a java class

get_doc
Handles token streams for javadoc comments and returns a Doc struct containing the information parsed from the javadoc comment.

get_method
Handles token streams for methods and returns a Method struct Containing the methods information from it's declaration

get_object
Handles token streams for object declarations and modifies the Class struct which is passed to the function.

get_var
Handles token streams for member variables and returns a Member struct Containing the member variable's data

lex_contents

match_params

parse_file
Root function of the module. Calls the lex and parse functions and returns a Class struct.

push_token